requirements.txt
# FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Machine Learning & Image Processing
numpy==1.24.3
opencv-python-headless==4.8.1.78  # Use headless version for server
Pillow==10.0.1
scikit-learn==1.3.2
joblib==1.3.2

# Background removal
rembg==2.0.38

# Optional: For environment variables
python-dotenv==1.0.0